safecp: Expose regular compiler features in safecpp
These are the regular compiler features that should be used by any
component that wants to claim to have good check on compiler warnings.
To not repeat the variable in every component, we put it in a single
place.
Because safecpp is a place to have tools to have safe code in C++,
this is a good location.

Issue: SWP-177717
GIT_ORIGIN_SPP_REV_ID: cb04e42a19a351ae38b05d030a55124fdc422968