Merge pull request #376 from swh/brem_remove_prd_assert_tx_log
lola: Change production assert in TransactionLog reset to debug

Since we currently have an issue with applications crashing due to a service being stop offered while a trace done callback is still pending, we remove this production assert and add a warning log.
In this case, the transaction would be cleaned up when re-offering the service, so there would be no harm to the application. Since the old callback may still be called due to an existing bug in the tracing library, it may result in the wrong slot being released in the new Skeleton. However, this would simply result in incorrect data being traced rather than a crash. Since this is anyway a short term fix, this is preferable to the existing production assert until the bug is fixed in the tracing library.
Issue: SWP-187920

Reviewed-by: manuelfehlhammer <Manuel.Fehlhammer@bmw.de>
GIT_ORIGIN_SPP_REV_ID: 0d44a932b708ce7920b68c77ead2cca86209d189