lola: Explicitly pass allocator to avoid implicit casting
Currently, the construct call for creating EventDataStorage will forward
number_of_slots (std::size_t) and MemoryResourceProxy* to the
DynamicArray. If SampleType is a bool, the MemoryResourceProxy* will be
implicitly cast to a bool and the DynamicArray constructor taking a
size, initial value (which is a bool) and allocator will be chosen. This
results in the PolymorphicOffsetPtrAllocator being default constructed,
since the MemoryResourceProxy* was used as the initial value instead of
being provided to the constructor of the PolymorphicOffsetPtrAllocator.
A default PolymorphicOffsetPtrAllocator will use a heap allocator.

Issue: SWP-107056
GIT_ORIGIN_SPP_REV_ID: 7147d2fa89638f02e9f8855a2f7f7ee676ad210c