bazel: Re-Enable all possible spawn-strategy
In a previous PR, we restricted the possible spwan strategies.
The idea was clear: We want to avoid an automatic fallback of bazel to
lesser spawn-strategies then `linux-sandbox`.

This comes with a ceaveat, we have some things that we do not want to
run inside the sandbox (or sometimes we want to use remote-execution).
Hardcoding on `linux-sandbox` basically avoids that some special tags
like `no-sandbox` will not work.

One way to overcome this is to specify a spawn strategy per mnemoic.
But for tests, they have all the same one: `TestRunner`.

The next alternative is `--strategy_regexp` but this does not work
for tests: https://github.com/bazelbuild/bazel/issues/18196

Thus, we took the decision to enable again all spawn_strategies.
The negative side-effect is that for some useres the sandbox will not
work out of the box, and they will have to take additional actions to
make this work - which they might only figure out through flaky tests.

The idea is to explain in some documentation sessions, how the
envrionement needs to be set up in this case.

But, the drawback of having the strategies hard-coded is to big for this
smaller problem for flaky tests.