memory::shared: Argue maybe-uninitalized warning
From GCC13 on, compiled with -O2, `maybe-uninitalized` is triggered.

We went through the code and as argued within the code, this is a false
positive. So, we disable the warning with an argumentation, in order to
enable compilation on newer GCC versions.

GIT_ORIGIN_SPP_REV_ID: c2e03e540f5e9dc187a80aa1e036183bda61a48c