lola: Always create shm
Remove optimization which doesn't create shm if there are no in args or return.
We currently rely on checking the shared memroy on the skeleton side to
know which methods are enabled. If we don't create a shm region when
there are no in args or return values, then we can't get this
information. Therefore, we create the shm region even in this case.

issue: SWP-211737
GIT_ORIGIN_SPP_REV_ID: e8d7954a126ff8a000f534ab035259df77d27124