Rename Release Verification Workflow to Coverage Report Workflow
We do not want to duplicate the logic of the workflows.
Since we have the logic except for coverage reports already as part
of other workflows we reduce this workflow to only create a coverage report.

All the other workflows we enable to also be triggered by a release creation.